+++ WireAuthentication/Tests/WireAuthenticationLogicTests/DetermineAuthMethodUseCaseTests.swift
@@ -227,62 +227,3 @@ private extension DomainRegistrationConfiguration {
     }
 
 }
-
-extension URLSession {
-
-    /// A url session powered by `URLProtocolMock`.
-    ///
-    /// Set `URLProtocolMock.mockHandler` with a mocking function to
-    /// control how request received by this URL session are handled.
-
-    static func mockURLSession() -> URLSession {
-        let config = URLSessionConfiguration.default
-        config.protocolClasses = [URLProtocolMock.self]
-        return URLSession(configuration: config)
-    }
-
-}
-
-final class URLProtocolMock: URLProtocol {
-
-    static var mockHandler: ((URLRequest) throws -> (Data?, URLResponse))?
-
-    override static func canInit(with request: URLRequest) -> Bool {
-        // This protocol can handle the request.
-        true
-    }
-
-    override static func canonicalRequest(for request: URLRequest) -> URLRequest {
-        // Let the request simply pass through.
-        request
-    }
-
-    override func startLoading() {
-        guard let mockHandler = Self.mockHandler else {
-            fatalError("no mock handler for `URLProtocolMock`")
-        }
-
-        let data: Data?
-        let response: URLResponse
-
-        do {
-            (data, response) = try mockHandler(request)
-        } catch {
-            client?.urlProtocol(self, didFailWithError: error)
-            return
-        }
-
-        client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-
-        if let data {
-            client?.urlProtocol(self, didLoad: data)
-        }
-
-        client?.urlProtocolDidFinishLoading(self)
-    }
-
-    override func stopLoading() {
-        // The request was cancelled or completed.
-    }
-
-}
